항해4주 : 회원가입, 로그인, 댓글 작성/수정/삭제 기능이 추가된 나만의 백엔드 서버

+++ HangHae_W4/src/main/java/com/sparta/week02/service/BoardService.java
@@ -1,10 +1,15 @@
 package com.sparta.week02.service;
 
 import com.sparta.week02.dto.BoardDto;
+import com.sparta.week02.dto.BoardResponseDto;
 import com.sparta.week02.dto.ResponseDto;
+import com.sparta.week02.handler.CustomException;
+import com.sparta.week02.handler.ErrorResponse;
 import com.sparta.week02.model.Board;
+import com.sparta.week02.model.Users;
 import com.sparta.week02.repository.BoardRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.sparta.week02.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -12,21 +17,32 @@ import java.util.List;
 import java.util.Map;
 
 @Service
+@RequiredArgsConstructor
 public class BoardService {
 
-    @Autowired
-    private BoardRepository boardRepository;
+
+    private final BoardRepository boardRepository;
+    private final UserRepository userRepository;
 
     @Transactional
-    public Board write(Map<String,String> param){
+    public BoardResponseDto write(Board board, Users user){
 //        System.out.println("pw : " + pw);
+        System.out.println("누구냐능 "+ user.getUsername());
         Board writeBoard = Board.builder()
-                .title(param.get("title"))
-                .content(param.get("content"))
-                .author(param.get("author"))
-                .password(param.get("password"))
+                .title(board.getTitle())
+                .content(board.getContent())
+                .author(user)
+                .password(board.getPassword())
                 .build();
-        return boardRepository.save(writeBoard);
+        System.out.println("저장됐냐능" + writeBoard);
+        boardRepository.save(writeBoard);
+        BoardResponseDto brd = BoardResponseDto.builder()
+                .id(writeBoard.getId())
+                .title(writeBoard.getTitle())
+                .content(writeBoard.getContent())
+                .author(writeBoard.getAuthor().getUsername())
+                .build();
+        return brd;
     }
 
     @Transactional(readOnly = true)
@@ -38,43 +54,49 @@ public class BoardService {
     @Transactional(readOnly = true)
     public Board getBoard(int id){
         return boardRepository.findById(id).orElseThrow(()->
-                new IllegalArgumentException("해당 게시글이 존재하지 않습니다."));
+                new CustomException(ErrorResponse.NOT_FOUND_BOARD));
     }
 
 
     // public int update(int id, Board board){
     // Dto 사용
     @Transactional
-    public Board updateBoard(int id, BoardDto boardDto){
+    public Board updateBoard(int id, BoardDto boardDto, Users user){
         // 영속화
         Board updateBoard = boardRepository.findById(id).orElseThrow(
-                ()-> new IllegalArgumentException("해당 게시글이 존재하지 않습니다.")
-        );
-        updateBoard.update(boardDto);
-        return updateBoard;
+                ()-> new CustomException(ErrorResponse.NOT_FOUND_BOARD));
+//        System.out.println("작성자 : " +updateBoard.getAuthor().getUsername() );
+//        System.out.println("로그인자 : " + user.getUsername());
+        if(updateBoard.getAuthor().getUsername().equals(user.getUsername())){
+            updateBoard.update(boardDto);
+            return updateBoard;
+        }else{
+           throw new CustomException(ErrorResponse.NOT_MATCH_USER);
+        }
     }
 
-    public int deleteBoard(int id) {
-        try{
+    public int deleteBoard(int id, Users user) {
+        Board updateBoard = boardRepository.findById(id).orElseThrow(
+                ()-> new CustomException(ErrorResponse.NOT_FOUND_BOARD));
+        if(updateBoard.getAuthor().getUsername().equals(user.getUsername())){
             boardRepository.deleteById(id);
-        }catch (Exception e){
-            e.printStackTrace();
+            return id;
+        }else{
+            throw new CustomException(ErrorResponse.NOT_MATCH_USER);
         }
-        return id;
     }
 
     public ResponseDto<String> certifyPw(int id, Map<String, String> map) {
         String pw = map.get("password");
 //        System.out.println("password = " + pw);
         Board board = boardRepository.findById(id).orElseThrow(
-                ()-> new IllegalArgumentException("해당 게시글이 존재하지 않습니다.")
-        );
+                ()-> new CustomException(ErrorResponse.NOT_FOUND_BOARD));
 
         if(board != null)
             if(board.getPassword().equals(pw))
-                return new ResponseDto<>(true, "비밀번호가 일치합니다.");
+                return new ResponseDto<>("SUCCESS", "비밀번호가 일치합니다.");
 
-        return new ResponseDto<>(false,"비밀번호가 일치하지 않습니다.");
+        return new ResponseDto<>("ERROR","비밀번호가 일치하지 않습니다.");
     }
 
 }
